fix: ignore updates without mtime change
Sometimes FS events are issued twice and the second time mtime does not
change. In that case, dispose callbacks are called, but module is not
actually reloaded since url does not change. Fixed by remembering last
module mtime and skipping reload if loaded version is already newer or
the same as updated one.